tests/wpt/tests/html/editing/dnd/images/009.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="../resources/test-helper.js"></script>
<head>
<title>Image drag and drop from iframe</title>
<style type="text/css">
div[ondragenter]
  {width:105px;
  min-height:105px;
  text-align:center;
  margin-top:20px;
  padding:10px;
  border:solid thin navy;}
iframe
  {width:150px;
  height:150px;
  border-style:none;}
</style>
<script type="application/ecmascript">
function addImage(event)
  {var c = document.createElement('img');
  c.setAttribute('src',event.dataTransfer.getData('text/uri-list').replace(/\r\n$/,''));
  document.querySelector('div').appendChild(c);}
</script>
</head>
<body>
<p><iframe src="helper-circle.xhtml">XHTML with image</iframe></p>
<p>Drag circle above to the box below. It should be copied to the box once you drop it there.</p>
<div
  ondragenter="event.preventDefault()"
  ondragover="return false"
/>
<script>
async function test() {
  await new Promise(loaded => window.addEventListener("load", loaded));
  const iframe = document.querySelector('iframe');
  const div = document.querySelector('div');
  function onDropCallBack(event) {
    addImage(event);
    var img = document.querySelector('img');
    assert_equals(img.src, event.dataTransfer.getData('text/uri-list').replace(/\r\n$/,''));
    return true;
  }
  dragDropTest(iframe, div, onDropCallBack, 'Dragging an image within an iframe to a div should copy it there');
};
test();
</script>
</body>
</html>
